pyproject.toml
[tool.poetry]
name = "to-chatgpt"
version = "0.1.0"
description = ""
authors = ["SimFG <1142838399@qq.com>"]
readme = "README.md"
packages = [{include = "to_chatgpt"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.97.0"
uvicorn = "^0.22.0"
httpx = "^0.24.1"
socksio = "^1.0.0"
pydantic = "^1.10.9"


[tool.poetry.group.new_bing.dependencies]
edgegpt = "^0.10.13"


[tool.poetry.group.bard.dependencies]
googlebard = "^1.3.2"


[tool.poetry.group.claude.dependencies]
anthropic = "^0.2.10"


[tool.poetry.group.cohere.dependencies]
cohere = "^4.11.2"


[tool.poetry.group.poe.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
